关键词加·，设置静态文件访问路径配置参数

+++ fun-read-server/src/main/java/com/bing/funread/server/controller/BaseController.java
@@ -11,6 +11,6 @@ import com.bing.funread.common.utils.TokenUtil;
 abstract class BaseController {
 
     Long getUserId() {
-        return TokenUtil.get(CommonConstant.TOKEN_USER_ID);
+        return ((Integer) TokenUtil.get(CommonConstant.TOKEN_USER_ID)).longValue();
     }
 }
